refactor passes test
