+++ src/constants/enums.ts
@@ -10,11 +10,12 @@ export enum Settings {
 
 export enum Commands {
   'resetColors' = 'peacock.resetColors',
-  'changeColor' = 'peacock.changeColor',
+  'enterColor' = 'peacock.enterColor',
   'changeColorToRandom' = 'peacock.changeColorToRandom',
   'changeColorToVueGreen' = 'peacock.changeColorToVueGreen',
   'changeColorToAngularRed' = 'peacock.changeColorToAngularRed',
-  'changeColorToReactBlue' = 'peacock.changeColorToReactBlue'
+  'changeColorToReactBlue' = 'peacock.changeColorToReactBlue',
+  'changeColorToPreferred' = 'peacock.changeColorToPreferred',
 }
 
 export enum ColorSettings {
