+++ src/constants/named-colors.ts
