+++ src/inputs.ts
@@ -1,0 +1,14 @@
+import * as vscode from 'vscode';
+import { BuiltInColors } from './constants/enums';
+
+export async function promptForColor() {
+  const options: vscode.InputBoxOptions = {
+    ignoreFocusOut: true,
+    placeHolder: BuiltInColors.Vue,
+    prompt:
+      'Enter a background color for the title bar in RGB hex format or a valid HTML color name',
+    value: BuiltInColors.Vue
+  };
+  const inputColor = await vscode.window.showInputBox(options);
+  return inputColor || '';
+}
